refactor: organiza estrutura do projeto
