+++ src/main/java/com/github/willcq97/scpods/api/controller/SubmissaoController.java
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.github.willcq97.scpods.api.dto.response.SubmissaoResponseDTO;
 import com.github.willcq97.scpods.api.dto.search.AcaoSearchDTO;
 import com.github.willcq97.scpods.api.dto.search.AcaoSearchOptionsDTO;
-import com.github.willcq97.scpods.domain.model.Acao;
+import com.github.willcq97.scpods.domain.model.entity.Acao;
 import com.github.willcq97.scpods.domain.service.AcaoService;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
