+++ src/main/java/com/github/willcq97/scpods/core/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.github.willcq97.scpods.config.swagger;
+package com.github.willcq97.scpods.core.config;
 
 import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
