+++ src/main/java/com/github/willcq97/scpods/domain/model/entity/Coordenador.java
@@ -1,4 +1,4 @@
-package com.github.willcq97.scpods.domain.model;
+package com.github.willcq97.scpods.domain.model.entity;
 
 import com.github.willcq97.scpods.domain.model.enums.TipoVinculoEnum;
