+++ src/main/java/com/github/willcq97/scpods/domain/model/entity/Local.java
@@ -1,4 +1,4 @@
-package com.github.willcq97.scpods.domain.model;
+package com.github.willcq97.scpods.domain.model.entity;
 
 import java.util.List;
 import java.util.Map;
