+++ src/main/java/com/github/willcq97/scpods/domain/model/entity/Unidade.java
@@ -1,4 +1,4 @@
-package com.github.willcq97.scpods.domain.model;
+package com.github.willcq97.scpods.domain.model.entity;
 
 import java.util.List;
 
@@ -11,6 +11,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -21,22 +22,23 @@ import lombok.Setter;
 @Getter
 @Setter
 @Entity
-@Table( name = "tb_lotacoes" )
-public class Lotacao {
+@Table( name = "tb_unidades" )
+public class Unidade {
 
     @Id
     private Long id;
 
     @NotBlank
-    private String descricao;
+    private String nome;
 
     @NotBlank
-    private String sigla;
+    private String codigo;
 
+    @NotNull
     @Enumerated( EnumType.STRING )
     private CampusEnum campus;
 
-    @OneToMany( mappedBy = "lotacao" )
-    private List<Acao> acoes;
+    @OneToMany( mappedBy = "unidade" )
+    private List<Local> locais;
 
 }
