+++ src/main/java/com/github/willcq97/scpods/domain/model/repository/AcaoRepository.java
@@ -1,4 +1,4 @@
-package com.github.willcq97.scpods.domain.repository;
+package com.github.willcq97.scpods.domain.model.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.ListCrudRepository;
 
-import com.github.willcq97.scpods.domain.model.Acao;
+import com.github.willcq97.scpods.domain.model.entity.Acao;
 
 public interface AcaoRepository extends ListCrudRepository<Acao, Long>, JpaSpecificationExecutor<Acao> {
