+++ src/main/java/com/github/willcq97/scpods/domain/model/repository/CoordenadorRepository.java
@@ -1,8 +1,8 @@
-package com.github.willcq97.scpods.domain.repository;
+package com.github.willcq97.scpods.domain.model.repository;
 
 import org.springframework.data.repository.ListCrudRepository;
 
-import com.github.willcq97.scpods.domain.model.Coordenador;
+import com.github.willcq97.scpods.domain.model.entity.Coordenador;
 
 public interface CoordenadorRepository extends ListCrudRepository<Coordenador, Long> {
