+++ src/main/java/com/github/willcq97/scpods/domain/model/repository/LocalRepository.java
@@ -1,8 +1,8 @@
-package com.github.willcq97.scpods.domain.repository;
+package com.github.willcq97.scpods.domain.model.repository;
 
 import org.springframework.data.repository.ListCrudRepository;
 
-import com.github.willcq97.scpods.domain.model.Local;
+import com.github.willcq97.scpods.domain.model.entity.Local;
 
 public interface LocalRepository extends ListCrudRepository<Local, Long> {
