+++ src/main/java/com/github/willcq97/scpods/domain/model/repository/LotacaoRepository.java
@@ -1,4 +1,4 @@
-package com.github.willcq97.scpods.domain.repository;
+package com.github.willcq97.scpods.domain.model.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.ListCrudRepository;
 
 import com.github.willcq97.scpods.api.dto.select.SelectModel;
-import com.github.willcq97.scpods.domain.model.Lotacao;
+import com.github.willcq97.scpods.domain.model.entity.Lotacao;
 import com.github.willcq97.scpods.domain.model.enums.CampusEnum;
 
 public interface LotacaoRepository extends ListCrudRepository<Lotacao, Long> {
