+++ src/main/java/com/github/willcq97/scpods/domain/model/repository/MetaRepository.java
@@ -1,10 +1,10 @@
-package com.github.willcq97.scpods.domain.repository;
+package com.github.willcq97.scpods.domain.model.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.ListCrudRepository;
 
-import com.github.willcq97.scpods.domain.model.Meta;
+import com.github.willcq97.scpods.domain.model.entity.Meta;
 
 public interface MetaRepository extends ListCrudRepository<Meta, Long> {
