+++ src/main/java/com/github/willcq97/scpods/domain/model/repository/ObjetivoRepository.java
@@ -1,4 +1,4 @@
-package com.github.willcq97.scpods.domain.repository;
+package com.github.willcq97.scpods.domain.model.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.ListCrudRepository;
 
-import com.github.willcq97.scpods.domain.model.Objetivo;
+import com.github.willcq97.scpods.domain.model.entity.Objetivo;
 
 public interface ObjetivoRepository extends ListCrudRepository<Objetivo, Long> {
