+++ src/main/java/com/github/willcq97/scpods/domain/service/ObjetivoService.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 
 import org.springframework.validation.annotation.Validated;
 
-import com.github.willcq97.scpods.domain.model.Meta;
-import com.github.willcq97.scpods.domain.model.Objetivo;
+import com.github.willcq97.scpods.domain.model.entity.Meta;
+import com.github.willcq97.scpods.domain.model.entity.Objetivo;
 
 @Validated
 public interface ObjetivoService {
