+++ src/main/java/com/github/willcq97/scpods/domain/service/impl/AcaoServiceImpl.java
@@ -12,19 +12,19 @@ import com.github.willcq97.scpods.api.dto.search.AcaoSearchDTO;
 import com.github.willcq97.scpods.api.dto.search.AcaoSearchOptionsDTO;
 import com.github.willcq97.scpods.domain.exception.BusinessException;
 import com.github.willcq97.scpods.domain.exception.EntityNotFoundException;
-import com.github.willcq97.scpods.domain.model.Acao;
-import com.github.willcq97.scpods.domain.model.Coordenador;
-import com.github.willcq97.scpods.domain.model.Local;
-import com.github.willcq97.scpods.domain.model.Lotacao;
-import com.github.willcq97.scpods.domain.model.Meta;
-import com.github.willcq97.scpods.domain.model.Objetivo;
-import com.github.willcq97.scpods.domain.model.Unidade;
+import com.github.willcq97.scpods.domain.model.entity.Acao;
+import com.github.willcq97.scpods.domain.model.entity.Coordenador;
+import com.github.willcq97.scpods.domain.model.entity.Local;
+import com.github.willcq97.scpods.domain.model.entity.Lotacao;
+import com.github.willcq97.scpods.domain.model.entity.Meta;
+import com.github.willcq97.scpods.domain.model.entity.Objetivo;
+import com.github.willcq97.scpods.domain.model.entity.Unidade;
 import com.github.willcq97.scpods.domain.model.enums.CampusEnum;
-import com.github.willcq97.scpods.domain.repository.AcaoRepository;
-import com.github.willcq97.scpods.domain.repository.LocalRepository;
-import com.github.willcq97.scpods.domain.repository.LotacaoRepository;
-import com.github.willcq97.scpods.domain.repository.MetaRepository;
-import com.github.willcq97.scpods.domain.repository.UnidadeRepository;
+import com.github.willcq97.scpods.domain.model.repository.AcaoRepository;
+import com.github.willcq97.scpods.domain.model.repository.LocalRepository;
+import com.github.willcq97.scpods.domain.model.repository.LotacaoRepository;
+import com.github.willcq97.scpods.domain.model.repository.MetaRepository;
+import com.github.willcq97.scpods.domain.model.repository.UnidadeRepository;
 import com.github.willcq97.scpods.domain.service.AcaoService;
 import com.github.willcq97.scpods.utils.SpecificationUtil;
