+++ src/main/java/com/github/willcq97/scpods/domain/service/impl/LotacaoServiceImpl.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.github.willcq97.scpods.api.dto.select.SelectModel;
-import com.github.willcq97.scpods.domain.repository.LotacaoRepository;
+import com.github.willcq97.scpods.domain.model.repository.LotacaoRepository;
 import com.github.willcq97.scpods.domain.service.LotacaoService;
 
 import lombok.AllArgsConstructor;
