+++ src/main/java/com/github/willcq97/scpods/domain/service/impl/ObjetivoServiceImpl.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
-import com.github.willcq97.scpods.domain.model.Meta;
-import com.github.willcq97.scpods.domain.model.Objetivo;
-import com.github.willcq97.scpods.domain.repository.MetaRepository;
-import com.github.willcq97.scpods.domain.repository.ObjetivoRepository;
+import com.github.willcq97.scpods.domain.model.entity.Meta;
+import com.github.willcq97.scpods.domain.model.entity.Objetivo;
+import com.github.willcq97.scpods.domain.model.repository.MetaRepository;
+import com.github.willcq97.scpods.domain.model.repository.ObjetivoRepository;
 import com.github.willcq97.scpods.domain.service.ObjetivoService;
 
 import lombok.AllArgsConstructor;
