+++ src/main/java/com/github/willcq97/scpods/domain/service/impl/UnidadeServiceImpl.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Service;
 import com.github.willcq97.scpods.api.dto.select.SelectModel;
 import com.github.willcq97.scpods.api.dto.select.SelectModelString;
 import com.github.willcq97.scpods.domain.exception.BusinessException;
-import com.github.willcq97.scpods.domain.model.Unidade;
+import com.github.willcq97.scpods.domain.model.entity.Unidade;
 import com.github.willcq97.scpods.domain.model.enums.CampusEnum;
-import com.github.willcq97.scpods.domain.repository.UnidadeRepository;
+import com.github.willcq97.scpods.domain.model.repository.UnidadeRepository;
 import com.github.willcq97.scpods.domain.service.UnidadeService;
 
 import jakarta.persistence.EntityNotFoundException;
